+++ bin/scrape_software_versions.py
@@ -56,9 +56,9 @@ for k, v in regexes.items():
 
 # Dump to YAML
 print ('''
-id: 'Chip-seq pipeline software versions'
+id: 'software_versions'
 section_name: 'Software Versions'
-section_href: 'https://gitlab.curie.fr/data-analysis/chip-seq/t'
+section_href: 'https://gitlab.curie.fr/data-analysis/chip-seq/'
 plot_type: 'html'
 description: 'are collected at run time from the software output.'
 data: |