+++ bin/stats2multiqc.sh
@@ -20,6 +20,7 @@ function help {
     exit;
 }
 
+is_pe=0
 while getopts "s:d:a:ph" OPT
 do
     case $OPT in
@@ -48,66 +49,55 @@ fi
 
 all_samples=$(awk -F, '{print $1}' $splan)
 
-echo -e "Sample_ID,Sample_name,Number_of_reads,Fragment_length,Number_of_aligned_reads,Percent_of_aligned_reads,Number_of_uniquely_mapped_reads,Percent_of_uniquely_mapped_reads,Number_of_multiple_mapped_reads,Percent_of_multiple_mapped_reads,Number_of_duplicates,Percent_of_duplicates,Normalized_strand_correlation,Relative_strand_correlation,Fraction_of_reads_in_peaks" > mqc.stats
+echo -e "Sample_ID,Sample_name,Number_of_reads,Fragment_length,Number_of_aligned_reads,Percent_of_aligned_reads,Number_of_hq_mapped_reads,Percent_of_hq_mapped_reads,Number_of_lq_mapped_reads,Percent_of_lq_mapped_reads,Number_of_duplicates,Percent_of_duplicates,Normalized_strand_correlation,Relative_strand_correlation,Fraction_of_reads_in_peaks" > mqc.stats
 
 for sample in $all_samples
 do
     #SAMPLE NAME
-    sname=$(grep "$sample" $splan | awk -F, '{print $2}')
+    sname=$(grep "$sample," $splan | awk -F, '{print $2}')
 
     #ALIGNMENT
     if [ $aligner == "bowtie2" ]; then
-	nb_reads=$(grep "reads;" mapping/${sample}_bowtie2.log | sed 's/ .*//')
-	if [[ $is_pe == "1" ]]; then
-           nb_uniq_reads=$(grep "concordantly exactly" mapping/${sample}_bowtie2.log | awk '{print $1}')
-           perc_uniq_reads=$(echo "${nb_uniq_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')                                                                                                   
-           nb_mult_reads=$(grep "concordantly >1" mapping/${sample}_bowtie2.log | awk '{print $1}')
-           perc_mult_reads=$(echo "${nb_mult_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')                                                                                                   
-           nb_mapped=$(($nb_uniq_reads + $nb_mult_reads))                                                                                                                                                   
-           perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')                                                                                                           
-	else 
-	    nb_uniq_reads=$(grep "exactly" mapping/${sample}_bowtie2.log | awk '{print $1}')
-	    perc_uniq_reads=$(echo "${nb_uniq_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-	    nb_mult_reads=$(grep ">1" mapping/${sample}_bowtie2.log | awk '{print $1}')
-	    perc_mult_reads=$(echo "${nb_mult_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-	    nb_mapped=$(($nb_uniq_reads + $nb_mult_reads))
-	    perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
+	nb_frag=$(grep "reads;" mapping/${sample}_bowtie2.log | sed 's/ .*//')
+	if [[ $is_pe == 1 ]]; then
+            nb_reads=$(( $nb_frag * 2 ))
+	else
+            nb_reads=$nb_frag
 	fi
     elif [ $aligner == "bwa-mem" ]; then
-	nb_unmapped=$(grep 'Unmapped' mapping/${sample}_bwa.log | awk -F "\t" '{print $2}')
-	nb_uniq_reads=$(grep 'Uniquely' mapping/${sample}_bwa.log | awk -F "\t" '{print $2}')
-	nb_mult_reads=$(grep 'Multi' mapping/${sample}_bwa.log | awk -F "\t" '{print $2}')
-	nb_reads=$(( $nb_unmapped + $nb_uniq_reads + $nb_mult_reads ))
-	perc_uniq_reads=$(echo "${nb_uniq_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-	perc_mult_reads=$(echo "${nb_mult_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-	nb_mapped=$(($nb_uniq_reads + $nb_mult_reads))
-	perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
+	# bwa.log file is in reads number (not pairs)
+	nb_reads=$(grep 'Total' mapping/${sample}_bwa.log | awk -F "\t" '{print $2}')
+	if [[ $is_pe == 1 ]]; then
+	    nb_frag=$(( $nb_reads / 2 ))
+	else
+	    nb_frag=$nb_reads
+	fi
+	tail -n +3 mapping/${sample}_bwa.log > mapping/${sample}_bwa.mqc
     elif [ $aligner == "star" ]; then
-	nb_reads=$(grep "Number of input reads" mapping/${sample}Log.final.out | cut -d"|" -f 2 | sed -e 's/\t//g')
-	nb_uniq_reads=$(grep "Uniquely.*number" mapping/${sample}Log.final.out | awk '{print $NF}')
-	perc_uniq_reads=$(echo "${nb_uniq_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-	nb_mult_reads=$(grep "Number.*multiple" mapping/${sample}Log.final.out | awk '{print $NF}')
-	perc_mult_reads=$(echo "${nb_mult_reads} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-	nb_mapped=$(($nb_uniq_reads + $nb_mult_reads))
-	perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
-    else
-	nb_reads='NA'
-	nb_uniq_reads='NA'
-	perc_uniq_reads='NA'
-	nb_mult_reads='NA'
-	perc_mult_reads='NA'
-	nb_mapped='NA'
-	perc_mapped='NA'
+	nb_frag=$(grep "Number of input reads" mapping/${sample}Log.final.out | cut -d"|" -f 2 | sed -e 's/\t//g')
+	if [[ $is_pe == 1 ]]; then
+            nb_reads=$(( $nb_frag * 2 ))
+	else
+            nb_reads=$nb_frag
+	fi
     fi
-  
+
+    #Mapping stats (always in reads - so must be converted for PE)
+    #These statistics are calculated after spike cleaning but before filtering
+    nb_mapped=$(awk -F, '$1=="Mapped"{print $2}' mapping/${sample}_mappingstats.mqc)
+    nb_mapped_hq=$(awk -F, '$1=="HighQual"{print $2}' mapping/${sample}_mappingstats.mqc)
+    nb_mapped_lq=$(awk -F, '$1=="LowQual"{print $2}' mapping/${sample}_mappingstats.mqc)
+    perc_mapped=$(echo "${nb_mapped} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
+    perc_mapped_hq=$(echo "${nb_mapped_hq} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
+    perc_mapped_lq=$(echo "${nb_mapped_lq} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
+
     #PICARD
     if [[ -e mapping/${sample}.MarkDuplicates.metrics.txt ]]; then
-	if [[ $is_pe == "1" ]]; then
-	    nb_dups=$(grep -a2 "## METRICS" mapping/${sample}.MarkDuplicates.metrics.txt | tail -1 | awk -F"\t" '{print $7}')
-	else
-	    nb_dups=$(grep -a2 "## METRICS" mapping/${sample}.MarkDuplicates.metrics.txt | tail -1 | awk -F"\t" '{print $6}')
-	fi
-	perc_dups=$(echo "${nb_dups} ${nb_reads}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
+	nb_dups_pair=$(grep -a2 "## METRICS" mapping/${sample}.MarkDuplicates.metrics.txt | tail -1 | awk -F"\t" '{print $7}')
+	nb_dups_single=$(grep -a2 "## METRICS" mapping/${sample}.MarkDuplicates.metrics.txt | tail -1 | awk -F"\t" '{print $6}')
+	nb_dups_optical=$(grep -a2 "## METRICS" mapping/${sample}.MarkDuplicates.metrics.txt | tail -1 | awk -F"\t" '{print $8}')
+	nb_dups=$(( $nb_dups_pair * 2 + $nb_dups_single + $nb_dups_optical ))
+	perc_dups=$(echo "${nb_dups} ${nb_mapped}" | awk ' { printf "%.*f",2,$1*100/$2 } ')
     else
 	nb_dups='NA'
 	perc_dups='NA'
@@ -138,6 +128,6 @@ do
     fi
 
     #To file
-    echo -e ${sample},${sname},${nb_reads},${frag_length},${nb_mapped},${perc_mapped},${nb_uniq_reads},${perc_uniq_reads},${nb_mult_reads},${perc_mult_reads},${nb_dups},${perc_dups},${nsc},${rsc},${frip} >> mqc.stats
+    echo -e ${sample},${sname},${nb_frag},${frag_length},${nb_mapped},${perc_mapped},${nb_mapped_hq},${perc_mapped_hq},${nb_mapped_lq},${perc_mapped_lq},${nb_dups},${perc_dups},${nsc},${rsc},${frip} >> mqc.stats
 done