+++ CHANGELOG.md
@@ -1,13 +1,36 @@
 ***********************************
-CAHNGES IN VERSION 1.0.2
+CHANGES IN VERSION 1.0.3
+
+NEW FEATURES
+
+   o Add --spikePercentFilter to fix the minimum percentage of reads which should align on spike-in genome (#23)
+   o Add --keepSingleton option
+   o Improve checkDesign function (#22)
+   o Add --noReadExtension parameters
+   o Add parameters for all mapper options
+
+BUG FIXES
+
+   o Fix bug for Epic2 without inputs (#47)
+   o Fix bugs in sample name (#39)
+   o Fix bug in spike-in reads cleaning
+   o Update BWA stats for PE data. Report pairs and not reads for mqc
+   o Dups % is calculated over the number of mapped reads (not total reads)
+
+***********************************
+CHANGES IN VERSION 1.0.2
 
 SIGNIFICANT USER-VISIBLE CHANGES
 
   o Add dmel6.22 annotation
+  o Report high/low quality reads instead of unique/multi reads
   o Update doc
+  o Use paired and single reads in duplicates level calculation
 
 BUG FIXES
 
+  o export plots in output documentation (#35)
+  o Add dmelr6.22 annotation genome
   o Fix bug in mapping stat for bowtie2 with paired-end data
 
 ***********************************